requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.1.1
django-bootstrap-v5==1.0.11
django-countries==7.5.1
django-crispy-forms==2.0
django-fontawesome-5==1.0.18
psycopg2-binary==2.9.6
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
